Add files via upload
In this assignment I solved a simple optimization problem for a function of one variable. Given a dataset of historical prices of a product from two suppliers, my task was to identify what share of the product I should buy from each of the suppliers to make the best possible investment in the future. Stating the problem mathematically, I constructed a target function to minimize, evaluate its minimum and investigated how its derivative is connected with the result.